API updates: connect() and more (#10)
## Summary

- Add connect() as a central way to connect to an organization.
- Document intents, transaction paths, transaction requests.
- Merge docs/README.md into README.md
- Move roadmap into ROADMAP.md.

## connect()

The connect() function is now the normal way to access an organization, and
handles both the connection and loading an organization. The idea here is to
optimize for the common scenario of building a thing that only needs to connect
to one organization.

It removes the need to have a “library object”, and it makes the connection to
the organization more central to the library, especially since the possibility
to create an organization has been removed for the time being.

Before:

const plumbery = new Plumbery()
const myorg = plumbery.organization('myorg')

After:

const myorg = await connect('myorg')

It also has been made asynchronous:

- It gives a chance to connectors to connect (we could have a standard
  ConnectorInterface#connect()).
- It could allow us to retrieve some information about the organization. This
  is something I was trying to avoid for performance reasons (the connection
  should be as fast as possible), but it could be kept minimal (e.g. only
  resolving the ENS name).

Which also means that it can now fail in different ways by throwing errors.

## Errors

I think errors could be put into two buckets: the normal exceptions (e.g. “the
org doesn’t seem to exist”), and the ones that shouldn’t happen (e.g.
“connect() requires at least one parameter”).

For the ones that shouldn’t happen, I think throw / catch is an obvious choice.

For other ones, it is common to try to avoid the verbosity of throw / catch for
other solutions. For example, if we take connect(), we could return null when
the organization doesn’t exist:

// Trying to avoid JS errors
import { connect } from 'plumbery'

const myorg = await connect('myorg.aragonid.eth')

if (myorg === null) {
  console.log('The org doesn’t seem to exist!')
}

// Using JS errors
import { connect } from 'plumbery'

try {
  const myorg = await connect('myorg.aragonid.eth')
} catch(err) {
  console.log('The org doesn’t seem to exist!')
}

Not throwing errors might look a little bit nicer − the first part doesn’t get
indented − but it comes with a few issues:

- Things can fail in only one way. For example, being able to make a
  distinction between a connection error and an organization not being found
  would require to introduce a new pattern.
- We lose the semantics of errors. A non-intercepted error having for message
  “The organization doesn’t seem to exist” is more useful than “Can’t access
  property “name” of null”. We don’t strictly need throw / catch to get
  the semantics of errors (e.g. Node callbacks or Promise#then()), but trying
  to avoid it would require a non-standard way to pass errors to the caller.
- It doesn’t interrupt the flow like non-catched errors do. Trying to avoid
  errors with e.g. replacing an object by null can create bugs difficult to
  understand if they cause indirect issues.

For these reasons, I think it could be reasonable to use throw / catch whenever
we want to communicate an exception, excluding obvious cases where the returned
value makes it really obvious (e.g. an empty array rather than a filled one is
generally enough to communicate “no results”).

Now I think the React API could do things a little bit differently, but I will
discuss this in another PR.

## IPFS

The connect() function now allows users to define an IPFS gateway resolver.

## Intent

Intent is an abstraction that represent anything that will end up in a
transaction. Every method resulting in a transaction now returns an intent
object.

It gives access to the different transaction paths, but also gives access to
the transaction(s) of the shortest one.

## TransactionPath

The transaction path is now represented using a simple data structure. It
contains the list of forwarders (apps), the last one (destination), and the
transaction(s) to sign.

We can have multiple transactions in some cases (“pretransaction”) that are not
up to the users of the library, so I think we need to communicate this idea of
having possibly multiple transactions. It will also allow us to do more things
in the future, like being able to assemble a basket of various transactions
before signing them all at once.

## TransactionRequest

There is no standard that I know of to represent a transaction request, but
Web3.js and ethers.js are both doing very similar things. The name “transaction
request” was inspired by the way ethers.js describes it, though they don’t have
a type for it.

The structure of this object is a combination of the two formats, and it also
has two methods: #toEthers() and #toWeb3(). Users of these libraries should
only have to call these to get the correct format.