Official subgraphs & Universal token manager (subgraph + connector) (#29)
Updates the token manager subgraph and connector to being univeral, so this is
pretty much the same as when we did it with Voting.

What was tricky in this one was on the subgraph, there was a non trivial
technical problem to be solved. Basically, TokenManager doesn't emit events
regarding to its managed token, so it was hard to pick up. It was solved by
introducing a hook which replaced the config in subgraphs. The voting subgraph
and the template subgraph were updated to use this as well.