Add calculateTransaction logic using ethers (#17)
Finally spend more time in the advanced calculate transaction path algorithm and most of the logic used by the simple path was the same.

Decide to not include the simple path as I'm not yet confident with the outcome and the way I encoding the actions. On the other hand, the calculatePath logic is working great and it should not be that hard to iterate on the simple version next.

I think we can merge this one and release the first alpha with what we currently have.